feat: club settings ui
